fix navigate to `about:blank/file://` page in the CDP mode (#7498)
1. Make the `isFileProtocol` condition less strong, since it's very
common to write file URI with just two slashes like `file://blabla`
instead of `file:///blabla`.
2. Add contextStorage restoring to the blank page to complete executing
function (COMMAND_EXECUTING_FLAG)
3. Use CDP redirect in the navigateTo command in case of the proxyless
mode and file protocol usage.

+++ src/client/driver/driver.js
@@ -117,6 +117,7 @@ import createErrorCtorCallback, {
 } from '../../shared/errors/selector-error-ctor-callback';
 import './command-executors/action-executor/actions-initializer';
 import { shouldSkipJsError } from './process-skip-js-errors';
+import isFileProtocol from '../../shared/utils/is-file-protocol';
 
 
 const settings       = hammerhead.settings;
@@ -1242,12 +1243,16 @@ export default class Driver extends serviceUtils.EventEmitter {
     _onNavigateToCommand (command) {
         this.contextStorage.setItem(this.COMMAND_EXECUTING_FLAG, true);
 
-        executeNavigateToCommand(command)
-            .then(driverStatus => {
-                this.contextStorage.setItem(this.COMMAND_EXECUTING_FLAG, false);
+        if (this.options.proxyless && isFileProtocol(command.url))
+            browser.redirectUsingCdp(command, hammerhead.createNativeXHR, this.communicationUrls.openFileProtocolUrl);
+        else {
+            executeNavigateToCommand(command)
+                .then(driverStatus => {
+                    this.contextStorage.setItem(this.COMMAND_EXECUTING_FLAG, false);
 
-                return this._onReady(driverStatus);
-            });
+                    return this._onReady(driverStatus);
+                });
+        }
     }
 
     _onGetProxyUrlCommand (command) {
